analytics/mongo.py
from pymongo import MongoClient
import os
from dotenv import load_dotenv
load_dotenv()

client = MongoClient(os.getenv("MONGO_URI"))
db = client["irctc_logs"]

search_logs_collection = db["train_search_logs"]